requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
Django==3.1.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
whitenoise==5.3.0
wrapt==1.12.1
